requirements.txt
stable-baselines3[extra]
gymnasium
notebook
matplotlib
pandas
numpy
scipy
gym
paho-mqtt